backend/requirements.txt
Django==5.0.6
djangorestframework==3.15.1
djangorestframework-simplejwt==5.3.1
django-cors-headers==4.3.1
psycopg[binary]==3.1.19
python-dotenv==1.0.1
django-environ==0.11.2
openai>=1.55,<2
pydantic==2.7.1
celery==5.4.0
redis==5.0.4
markdown==3.6
weasyprint==62.1
python-docx==1.1.2
gunicorn==22.0.0
